third_party/blink/web_tests/svg/dynamic-updates/SVGRadialGradientElement-dom-r-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'r' attribute of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.getAttribute('r') is "125%"
PASS radialGradientElement.getAttribute('r') is "50%"
PASS successfullyParsed is true

TEST COMPLETE
